refactor(app): extract prompt/minibuffer lifecycle (phase 6d)
Move the prompt orchestrator cluster (start_prompt, start_search_prompt,
start_prompt_with_suggestions, start_prompt_with_initial_text,
start_quick_open, build_quick_open_context, update_quick_open_suggestions,
cancel_search_prompt_if_active, prefill_open_file_prompt,
init_file_open_state, init_folder_open_state, change_working_dir,
load_file_open_directory, handle_file_open_directory_loaded,
load_file_open_shortcuts_async, handle_file_open_shortcuts_loaded,
cancel_prompt, handle_prompt_scroll, confirm_prompt, is_prompting,
get_or_create_prompt_history, get_prompt_history,
prompt_type_to_history_key, editor_mode, command_registry,
plugin_manager(_mut), file_explorer_is_focused, prompt_input,
has_active_selection, prompt_mut, set_status_message, get_status_message,
get_plugin_errors, clear_plugin_errors, update_prompt_suggestions) out
of mod.rs into a new app/prompt_lifecycle.rs sibling. ~824 lines.

mod.rs goes from 5705 -> 4882 lines (-823). 2026 tests still pass.
